Add functional title_case.cpp with 2 known bugs
This program works mostly as intended, however there are two bugs:

- When the final word in the input has one or more trailing spaces, it
will not be capitalized. The bug can be found on line 40. The program
looks for spaces to indicate that there will be a new word, so it
mistakenly assumes a new word is coming when the input has trailing
spaces even when there is not.

- When one of the non-capitalized words is followed by a special
character (e.g. ',') it will be capitalized. The bug can be found on
line 50. The program looks for an exact match of these words to
determine not to capitalize them, so if a special character follows it
preceding the space, it will mistakenly read this as a different word.