Add functional six_degrees.cpp
This program works as intended, but it may be possible to improve the
efficiency and readability in future updates. Also, this version can
frustrate the user on input: matrices of larger sizes quickly become
unwieldy to manage and are prone to accidental input mistakes that
necessitate starting over. Reading the matrix from a file would prevent
this issue.